glTF-Exporter gehärtet: IBMs neu berechnet und saniert, byteStride entfernt, Joints/Weights normalisiert/gesäubert, Mesh-Node als Scene-Root gesetzt, Identitätsmatrizen weggelassen. EAN-Analyse-Tool hinzugefügt (ean_dump.py) mit Heuristiken für Keyframes/Tracks und JSON-Dumps. DFF→glTF-Exports erneut ausgeführt für a000000a (Validator jetzt clean).

+++ Archlord-AIO/libs/rw_dff/src/tree.rs
@@ -20,7 +20,7 @@ pub struct RwChunkNode {
 impl RwChunkNode {
     /// Returns the total size of this chunk on disk (header + payload).
     pub fn total_size(&self) -> u64 {
-        (self.payload_end - self.header_off) as u64
+        self.payload_end - self.header_off
     }
 }
